requirements.txt
pandas
transformers
faiss-cpu
streamlit
openai
nltk